Major

+++ Models/SetTags.cs
@@ -9,7 +9,7 @@ namespace APMD.Data.Models
     public class SetTags
     {
         public int PK_SET_TAG_ID { get; set; }
-        public int FK_SET_ID { get; set; }
+        public long FK_SET_ID { get; set; }
         public int FK_TAG_ID { get; set; }
     }
 }
